refactor(shipping): Phase 2b move domain to Domain/
- Relocate 9 Models/* types to Domain/ with namespace
  Shipping.Service.Domain; rebase 47 consumer using statements;
  no behavior change.
- Models/ folder deleted; build + Shipping.Tests green
  (107 pass, 7 Phase-10 skips).
