+++ shipping-microservice/Shipping.Service/Carriers/RateShoppingService.cs
@@ -1,4 +1,4 @@
-using Shipping.Service.Models;
+using Shipping.Service.Domain;
 
 namespace Shipping.Service.Carriers;
