+++ shipping-microservice/Shipping.Service/Domain/Money.cs
@@ -1,4 +1,4 @@
-namespace Shipping.Service.Models;
+namespace Shipping.Service.Domain;
 
 public record Money(decimal Amount, string Currency)
 {
