+++ shipping-microservice/Shipping.Service/Domain/ShipmentStatusHistoryEntry.cs
@@ -1,4 +1,4 @@
-namespace Shipping.Service.Models;
+namespace Shipping.Service.Domain;
 
 internal class ShipmentStatusHistoryEntry
 {
