+++ shipping-microservice/Shipping.Service/Domain/ShippingAddress.cs
@@ -1,4 +1,4 @@
-namespace Shipping.Service.Models;
+namespace Shipping.Service.Domain;
 
 public record ShippingAddress(
     string Recipient,
