+++ shipping-microservice/Shipping.Service/Infrastructure/Data/EntityFramework/ShipmentLineConfiguration.cs
@@ -1,7 +1,7 @@
 using ECommerce.Shared.Qa;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Shipping.Service.Models;
+using Shipping.Service.Domain;
 
 namespace Shipping.Service.Infrastructure.Data.EntityFramework;
