+++ shipping-microservice/Shipping.Service/Infrastructure/Data/IShipmentStore.cs
@@ -1,4 +1,4 @@
-using Shipping.Service.Models;
+using Shipping.Service.Domain;
 
 namespace Shipping.Service.Infrastructure.Data;
