+++ shipping-microservice/Shipping.Service/IntegrationEvents/EventHandlers/CancelShipmentCommandHandler.cs
@@ -3,8 +3,8 @@ using ECommerce.Shared.Infrastructure.EventBus.Abstractions;
 using ECommerce.Shared.Infrastructure.Outbox;
 using ECommerce.Shared.IntegrationEvents.Commands;
 using Shipping.Service.Contracts.Integration;
+using Shipping.Service.Domain;
 using Shipping.Service.Infrastructure.Data;
-using Shipping.Service.Models;
 
 namespace Shipping.Service.IntegrationEvents.EventHandlers;
