+++ shipping-microservice/Shipping.Tests/Api/CreateShipmentCommandHandlerTests.cs
@@ -3,8 +3,8 @@ using ECommerce.Shared.Infrastructure.Outbox;
 using ECommerce.Shared.IntegrationEvents.Commands;
 using Microsoft.Extensions.DependencyInjection;
 using Shipping.Service.Contracts.Integration;
+using Shipping.Service.Domain;
 using Shipping.Service.IntegrationEvents.EventHandlers;
-using Shipping.Service.Models;
 
 namespace Shipping.Tests.Api;
