+++ shipping-microservice/Shipping.Tests/Api/ShipmentDispatchEndpointsTests.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Shipping.Service.ApiModels;
 using Shipping.Service.Carriers;
 using Shipping.Service.Contracts.Integration;
-using Shipping.Service.Models;
+using Shipping.Service.Domain;
 using Shipping.Tests.Authentication;
 
 namespace Shipping.Tests.Api;
