+++ shipping-microservice/Shipping.Tests/Api/ShipmentTerminalTransitionTests.cs
@@ -4,7 +4,7 @@ using ECommerce.Shared.Infrastructure.Outbox;
 using Microsoft.Extensions.DependencyInjection;
 using Shipping.Service.ApiModels;
 using Shipping.Service.Contracts.Integration;
-using Shipping.Service.Models;
+using Shipping.Service.Domain;
 using Shipping.Tests.Authentication;
 
 namespace Shipping.Tests.Api;
