+++ shipping-microservice/Shipping.Tests/Carriers/CarrierGatewayContractTests.cs
@@ -1,5 +1,5 @@
 using Shipping.Service.Carriers;
-using Shipping.Service.Models;
+using Shipping.Service.Domain;
 
 namespace Shipping.Tests.Carriers;
