+++ shipping-microservice/Shipping.Tests/Domain/ShipmentTests.cs
@@ -1,4 +1,4 @@
-using Shipping.Service.Models;
+using Shipping.Service.Domain;
 
 namespace Shipping.Tests.Domain;
